+++ xmake/core/sandbox/modules/utils.lua
@@ -39,7 +39,6 @@ sandbox_utils.confirm = utils.confirm
 sandbox_utils.error   = utils.error
 sandbox_utils.warning = utils.warning
 sandbox_utils.trycall = utils.trycall
-sandbox_utils.ifelse  = utils.ifelse
 
 -- print each arguments
 function sandbox_utils._print(...)
@@ -60,25 +59,20 @@ end
 -- print multi-variables with raw lua action
 --
 function sandbox_utils.print(format, ...)
-
-    -- print format string
     if type(format) == "string" and format:find("%", 1, true) then
         local args = {...}
-        try
-        {
+        try {
             function ()
                 local message = vformat(format, table.unpack(args))
                 utils._print(message)
                 log:printv(message)
             end,
-            catch
-            {
+            catch {
                 function (errors)
                     sandbox_utils._print(format, table.unpack(args))
                 end
             }
         }
-
     else
         sandbox_utils._print(format, ...)
     end