+++ xmake/modules/package/tools/cmake.lua
@@ -20,6 +20,7 @@
 
 -- imports
 import("core.base.option")
+import("core.base.semver")
 import("core.tool.toolchain")
 import("core.project.config")
 import("core.tool.linker")
@@ -86,7 +87,7 @@ end
 
 -- get msvc
 function _get_msvc(package)
-    local msvc = toolchain.load("msvc", {plat = package:plat(), arch = package:arch()})
+    local msvc = package:toolchain("msvc") or toolchain.load("msvc", {plat = package:plat(), arch = package:arch()})
     assert(msvc:check(), "vs not found!") -- we need to check vs envs if it has been not checked yet
     return msvc
 end
@@ -274,6 +275,19 @@ function _get_shflags(package, opt)
     end
 end
 
+-- get cmake version
+function _get_cmake_version()
+    local cmake_version = _g.cmake_version
+    if not cmake_version then
+        local cmake = find_tool("cmake", {version = true})
+        if cmake and cmake.version then
+            cmake_version = semver.new(cmake.version)
+        end
+        _g.cmake_version = cmake_version
+    end
+    return cmake_version
+end
+
 -- get vs toolset
 function _get_vs_toolset(package)
     local toolset_ver = nil
@@ -282,6 +296,14 @@ function _get_vs_toolset(package)
         local verinfo = vs_toolset:split('%.')
         if #verinfo >= 2 then
             toolset_ver = "v" .. verinfo[1] .. (verinfo[2]:sub(1, 1) or "0")
+        end
+    end
+    -- cmake does not support vs toolset v144 below 3.29.0, we can only use v143
+    -- @see https://github.com/xmake-io/xmake/issues/4772
+    if toolset_ver and toolset_ver >= "v144" then
+        local cmake_version = _get_cmake_version()
+        if cmake_version and cmake_version:le("3.29.0") then
+            toolset_ver = "v143"
         end
     end
     return toolset_ver
@@ -382,9 +404,9 @@ function _get_configs_for_windows(package, configs, opt)
 end
 
 -- get configs for android
+-- https://developer.android.google.cn/ndk/guides/cmake
 function _get_configs_for_android(package, configs, opt)
-
-    -- https://developer.android.google.cn/ndk/guides/cmake
+    opt = opt or {}
     local ndk = get_config("ndk")
     if ndk and os.isdir(ndk) then
         local ndk_sdkver = get_config("ndk_sdkver")
@@ -398,7 +420,7 @@ function _get_configs_for_android(package, configs, opt)
         if ndk_cxxstl then
             table.insert(configs, "-DANDROID_STL=" .. ndk_cxxstl)
         end
-        if is_host("windows") then
+        if is_host("windows") and opt.cmake_generator ~= "Ninja" then
             local make = path.join(ndk, "prebuilt", "windows-x86_64", "bin", "make.exe")
             if os.isfile(make) then
                 table.insert(configs, "-DCMAKE_MAKE_PROGRAM=" .. make)
@@ -473,7 +495,7 @@ function _get_configs_for_mingw(package, configs, opt)
     -- Avoid cmake to add the flags -search_paths_first and -headerpad_max_install_names on macOS
     envs.HAVE_FLAG_SEARCH_PATHS_FIRST = "0"
     -- CMAKE_MAKE_PROGRAM may be required for some CMakeLists.txt (libcurl)
-    if is_subhost("windows") then
+    if is_subhost("windows") and opt.cmake_generator ~= "Ninja" then
         local mingw = assert(package:build_getenv("mingw") or package:build_getenv("sdk"), "mingw not found!")
         envs.CMAKE_MAKE_PROGRAM = path.join(mingw, "bin", "mingw32-make.exe")
     end
@@ -485,6 +507,7 @@ end
 
 -- get configs for wasm
 function _get_configs_for_wasm(package, configs, opt)
+    opt = opt or {}
     local emsdk = find_emsdk()
     assert(emsdk and emsdk.emscripten, "emscripten not found!")
     local emscripten_cmakefile = find_file("Emscripten.cmake", path.join(emsdk.emscripten, "cmake/Modules/Platform"))
@@ -1062,3 +1085,4 @@ function install(package, configs, opt)
     end
     os.cd(oldir)
 end
+