+++ xmake/rules/c++/modules/modules_support/gcc/builder.lua
@@ -183,7 +183,7 @@ function make_module_buildjobs(target, batchjobs, job_name, deps, opt)
         name = job_name,
         deps = table.join(target:name() .. "_populate_module_map", deps),
         sourcefile = opt.cppfile,
-        job = batchjobs:newjob(name or opt.cppfile, function(index, total)
+        job = batchjobs:newjob(name or opt.cppfile, function(index, total, jobopt)
             local mapped_bmi
             if provide and compiler_support.memcache():get2(target:name() .. name, "reuse") then
                 if not target:is_binary() then
@@ -214,23 +214,24 @@ function make_module_buildjobs(target, batchjobs, job_name, deps, opt)
                     local fileconfig = target:fileconfig(opt.cppfile)
                     local public = fileconfig and fileconfig.public
                     local external = fileconfig and fileconfig.external
+                    local private_dep = fileconfig and fileconfig.private_dep
                     local bmifile = mapped_bmi or bmifile
                     local flags = {"-x", "c++"}
                     local sourcefile
                     if target:is_binary() then
                         if mapped_bmi then
-                            progress.show((index * 100) / total, "${color.build.target}<%s> ${clear}${color.build.object}compiling.objectfile.$(mode) %s", target:name(), name or opt.cppfile)
+                            progress.show(jobopt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.objectfile.$(mode) %s", target:name(), name or opt.cppfile)
                             sourcefile = bmifile
                         else
-                            progress.show((index * 100) / total, "${color.build.target}<%s> ${clear}${color.build.object}compiling.module.$(mode) %s", target:name(), name or opt.cppfile)
+                            progress.show(jobopt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.module.$(mode) %s", target:name(), name or opt.cppfile)
                             sourcefile = opt.cppfile
                         end
                     else
-                        if not public and not external then
-                            progress.show((index * 100) / total, "${color.build.target}<%s> ${clear}${color.build.object}compiling.module.$(mode) %s", target:name(), name or opt.cppfile)
+                        if (not public and not external) or (external and private_dep) then
+                            progress.show(jobopt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.module.$(mode) %s", target:name(), name or opt.cppfile)
                             sourcefile = opt.cppfile
                         else
-                            progress.show((index * 100) / total, "${color.build.target}<%s> ${clear}${color.build.object}compiling.bmi.$(mode) %s", target:name(), name or opt.cppfile)
+                            progress.show(jobopt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.bmi.$(mode) %s", target:name(), name or opt.cppfile)
                             local module_onlyflag = compiler_support.get_moduleonlyflag(target)
                             table.insert(flags, module_onlyflag)
                             sourcefile = opt.cppfile
@@ -272,50 +273,42 @@ function make_module_buildcmds(target, batchcmds, opt)
         target:fileconfig_add(opt.cppfile, {force = {cxxflags = {module_mapperflag .. module_mapper}}})
     end
 
-    local build = should_build(target, opt.cppfile, bmifile, {name = name, objectfile = opt.objectfile, requires = opt.module.requires})
-
-    -- needed to detect rebuild of dependencies
-    if provide and build then
-        mark_build(target, name)
-    end
-
-    if build then
-        -- compile if it's a named module
-        if provide or compiler_support.has_module_extension(opt.cppfile) then
-            batchcmds:mkdir(path.directory(opt.objectfile))
-            local fileconfig = target:fileconfig(opt.cppfile)
-            local public = fileconfig and fileconfig.public
-            local external = fileconfig and fileconfig.external
-            local bmifile = mapped_bmi or bmifile
-            local flags = {"-x", "c++"}
-            local sourcefile
-            if target:is_binary() then
-                if mapped_bmi then
-                    batchcmds:show_progress(opt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.objectfile.$(mode) %s", target:name(), name or opt.cppfile)
-                    sourcefile = bmifile
-                else
-                    batchcmds:show_progress(opt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.module.$(mode) %s", target:name(), name or opt.cppfile)
-                    sourcefile = opt.cppfile
-                end
+    -- compile if it's a named module
+    if provide or compiler_support.has_module_extension(opt.cppfile) then
+        batchcmds:mkdir(path.directory(opt.objectfile))
+        local fileconfig = target:fileconfig(opt.cppfile)
+        local public = fileconfig and fileconfig.public
+        local external = fileconfig and fileconfig.external
+        local private_dep = fileconfig and fileconfig.private_dep
+        local bmifile = mapped_bmi or bmifile
+        local flags = {"-x", "c++"}
+        local sourcefile
+        if target:is_binary() then
+            if mapped_bmi then
+                batchcmds:show_progress(opt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.objectfile.$(mode) %s", target:name(), name or opt.cppfile)
+                sourcefile = bmifile
             else
-                if not public and not external then
-                    batchcmds:show_progress(opt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.module.$(mode) %s", target:name(), name or opt.cppfile)
-                    sourcefile = opt.cppfile
-                else
-                    batchcmds:show_progress(opt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.bmi.$(mode) %s", target:name(), name or opt.cppfile)
-                    local module_onlyflag = compiler_support.get_moduleonlyflag(target)
-                    table.insert(flags, module_onlyflag)
-                    sourcefile = opt.cppfile
-                end
+                batchcmds:show_progress(opt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.module.$(mode) %s", target:name(), name or opt.cppfile)
+                sourcefile = opt.cppfile
             end
-            if option.get("diagnosis") then
-                batchcmds:print("mapper file: %s", io.readfile(module_mapper))
-            end
-            _batchcmds_compile(batchcmds, target, flags, sourcefile, opt.objectfile)
-            batchcmds:rm(module_mapper)
         else
-            batchcmds:rm(opt.objectfile) -- force rebuild for .cpp files
+            if (not public and not external) or (external and private_dep) then
+                batchcmds:show_progress(opt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.module.$(mode) %s", target:name(), name or opt.cppfile)
+                sourcefile = opt.cppfile
+            else
+                batchcmds:show_progress(opt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.bmi.$(mode) %s", target:name(), name or opt.cppfile)
+                local module_onlyflag = compiler_support.get_moduleonlyflag(target)
+                table.insert(flags, module_onlyflag)
+                sourcefile = opt.cppfile
+            end
         end
+        if option.get("diagnosis") then
+            batchcmds:print("mapper file: %s", io.readfile(module_mapper))
+        end
+        _batchcmds_compile(batchcmds, target, flags, sourcefile, opt.objectfile)
+        batchcmds:rm(module_mapper)
+    else
+        batchcmds:rm(opt.objectfile) -- force rebuild for .cpp files
     end
     batchcmds:add_depfiles(opt.cppfile)
     return os.mtime(opt.objectfile)
@@ -331,7 +324,7 @@ function make_headerunit_buildjobs(target, job_name, batchjobs, headerunit, bmif
         return {
             name = job_name,
             sourcefile = headerunit.path,
-            job = batchjobs:newjob(job_name, function(index, total)
+            job = batchjobs:newjob(job_name, function(index, total, jobopt)
                 if not os.isdir(outputdir) then
                     os.mkdir(outputdir)
                 end
@@ -346,7 +339,7 @@ function make_headerunit_buildjobs(target, job_name, batchjobs, headerunit, bmif
 
                 if opt.build then
                     local headerunit_mapper = _generate_headerunit_modulemapper_file({name = path.normalize(headerunit.path), bmifile = bmifile})
-                    progress.show((index * 100) / total, "${color.build.target}<%s> ${clear}${color.build.object}compiling.headerunit.$(mode) %s", target:name(), headerunit.name)
+                    progress.show(jobopt.progress, "${color.build.target}<%s> ${clear}${color.build.object}compiling.headerunit.$(mode) %s", target:name(), headerunit.name)
                     if option.get("diagnosis") then
                         print("mapper file:\n%s", io.readfile(headerunit_mapper))
                     end